Convert internal immediate through i64 (#546)
_Follow-up to #536 _

Consolidate the processing of immediate values into a single `i64`
instead of previous `u32` and `bool`.

---------

+++ ceno_zkvm/src/instructions/riscv/arith_imm.rs
@@ -62,7 +62,10 @@ impl<E: ExtensionField> Instruction<E> for AddiInstruction<E> {
         step: &StepRecord,
     ) -> Result<(), ZKVMError> {
         let rs1_read = Value::new_unchecked(step.rs1().unwrap().value);
-        let imm = Value::new(InsnRecord::imm_internal(&step.insn()), lk_multiplicity);
+        let imm = Value::new(
+            InsnRecord::imm_internal(&step.insn()) as u32,
+            lk_multiplicity,
+        );
 
         let result = rs1_read.add(&imm, lk_multiplicity, true);
